Created a non-root user to fix x11 forwarding
Because XDG_RUNTIME_DIR was owned by uid 1000 (non-root user) on the
host (and this is probably is true for all Linux distros) but by uid 0
(root user) inside of the container and this was breaking x11
forwarding.

Creating a non-root user and setting it as the standard container user
fixed this issue entirely.

modified:   Dockerfile